% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{classify_gene_layer}
\alias{classify_gene_layer}
\title{Assign a gene to a hierarchy layer}
\usage{
classify_gene_layer(gene_name)
}
\arguments{
\item{gene_name}{Gene symbol.}
}
\value{
\code{"Clock"}, \code{"Target"} or \code{"Background"}.
}
\description{
Assign a gene to a hierarchy layer
}
\examples{
classify_gene_layer("Per2")
classify_gene_layer("Actb")

}
