% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{gearbox_gap}
\alias{gearbox_gap}
\title{Gearbox gap between two sets of eigenvalues}
\usage{
gearbox_gap(clock_eigenvalues, target_eigenvalues)
}
\arguments{
\item{clock_eigenvalues, target_eigenvalues}{Numeric vectors of
\eqn{|\lambda|}.}
}
\value{
\code{median(clock) - median(target)}, or \code{0} if either set is
  empty.
}
\description{
Gearbox gap between two sets of eigenvalues
}
\examples{
gearbox_gap(c(0.72, 0.70), c(0.61, 0.59))

}
